+++ api_tests/server_api_test.py
@@ -9,11 +9,17 @@ def test_get_info(base_url, headers):
 
     assert resp.status_code == 200
     data = json.loads(resp.content)
+<<<<<<< HEAD
     assert data['version'] == '0.4.3'
+=======
+    assert data['version'] == '0.5.0'
+>>>>>>> main
     assert int(data['bucket_count']) > 0
     assert int(data['uptime']) >= 0
     assert int(data['latest_record']) >= 0
     assert int(data['oldest_record']) >= 0
+
+    assert data['defaults']['bucket'] == {'max_block_size': '67108864', 'quota_size': '0', 'quota_type': 'NONE'}
 
 
 def test_get_list_of_buckets(base_url, headers):
@@ -33,10 +39,6 @@ def test_get_list_of_buckets(base_url, headers):
 
 def test_get_wrong_path(base_url, headers):
     """Should return 404"""
-
-    resp = requests.get(f'{base_url}/', headers=headers)
-    assert resp.status_code == 404
-
     resp = requests.get(f'{base_url}/NOTEXIST', headers=headers)
     assert resp.status_code == 404