+++ src/reduct/storage/storage.cc
@@ -47,6 +47,11 @@ class Storage : public IStorage {
 
     return Run<Callback::Result>([this] {
       using proto::api::ServerInfo;
+<<<<<<< HEAD
+=======
+      using proto::api::Defaults;
+      using proto::api::BucketSettings;
+>>>>>>> main
 
       size_t usage = 0;
       uint64_t oldest_ts = std::numeric_limits<uint64_t>::max();
@@ -68,8 +73,8 @@ class Storage : public IStorage {
       info.set_oldest_record(oldest_ts);
       info.set_latest_record(latest_ts);
 
-      Callback::Response resp{.info = std::move(info)};
-      return Callback::Result{std::move(resp), Error{}};
+      *info.mutable_defaults()->mutable_bucket() = IBucket::GetDefaults();
+      return Callback::Result{std::move(info), Error{}};
     });
   }